+++ src/main/java/ministryofeducation/sideprojectspring/personnel/application/PersonnelService.java
@@ -4,18 +4,18 @@ import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import ministryofeducation.sideprojectspring.personnel.domain.Personnel;
-import ministryofeducation.sideprojectspring.personnel.infrastructure.PersonnelListRepository;
+import ministryofeducation.sideprojectspring.personnel.infrastructure.PersonnelRepository;
 import ministryofeducation.sideprojectspring.personnel.presentation.dto.response.PersonnelListDto;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class PersonnelListService {
+public class PersonnelService {
 
-    private final PersonnelListRepository personnelListRepository;
+    private final PersonnelRepository personnelRepository;
 
     public List<PersonnelListDto> personnelList(){
-        List<PersonnelListDto> personnelListResponse = personnelListRepository.findAll().stream()
+        List<PersonnelListDto> personnelListResponse = personnelRepository.findAll().stream()
             .map(this::getPersonnelListDto)
             .collect(Collectors.toList());
