+++ src/main/java/ministryofeducation/sideprojectspring/personnel/infrastructure/PersonnelRepository.java
@@ -4,6 +4,6 @@ import java.util.List;
 import ministryofeducation.sideprojectspring.personnel.domain.Personnel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonnelListRepository extends JpaRepository<Personnel, Long> {
+public interface PersonnelRepository extends JpaRepository<Personnel, Long> {
     List<Personnel> findAll();
 }
