+++ src/main/java/ministryofeducation/sideprojectspring/personnel/presentation/PersonnelController.java
@@ -2,7 +2,7 @@ package ministryofeducation.sideprojectspring.personnel.presentation;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import ministryofeducation.sideprojectspring.personnel.application.PersonnelListService;
+import ministryofeducation.sideprojectspring.personnel.application.PersonnelService;
 import ministryofeducation.sideprojectspring.personnel.presentation.dto.response.PersonnelListDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,14 +10,15 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-public class PersonnelListController {
+public class PersonnelController {
 
-    private final PersonnelListService personnelListService;
+    private final PersonnelService personnelService;
 
     @GetMapping("/list")
     public ResponseEntity<List<PersonnelListDto>> personnelList(){
-        List<PersonnelListDto> responseDto = personnelListService.personnelList();
+        List<PersonnelListDto> responseDto = personnelService.personnelList();
 
         return ResponseEntity.ok(responseDto);
     }
+
 }
