+++ src/test/java/ministryofeducation/sideprojectspring/Personnel_list/application/PersonnelServiceTest.java
@@ -4,27 +4,18 @@ import static ministryofeducation.sideprojectspring.factory.PersonnelFactory.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
-import java.time.LocalDate;
 import java.util.List;
 
-import jakarta.persistence.EntityManager;
-import ministryofeducation.sideprojectspring.factory.PersonnelFactory;
-import ministryofeducation.sideprojectspring.personnel_list.application.PersonnelListService;
-import ministryofeducation.sideprojectspring.personnel_list.domain.Personnel;
-import ministryofeducation.sideprojectspring.personnel_list.infrastructure.PersonnelListRepository;
-import ministryofeducation.sideprojectspring.personnel_list.presentation.dto.response.PersonnelListDto;
-import org.junit.jupiter.api.BeforeEach;
+import ministryofeducation.sideprojectspring.personnel.application.PersonnelListService;
+import ministryofeducation.sideprojectspring.personnel.infrastructure.PersonnelListRepository;
+import ministryofeducation.sideprojectspring.personnel.presentation.dto.response.PersonnelListDto;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
