+++ src/test/java/ministryofeducation/sideprojectspring/Personnel_list/presentation/PersonnelControllerTest.java
@@ -2,17 +2,13 @@ package ministryofeducation.sideprojectspring.Personnel_list.presentation;
 
 import static ministryofeducation.sideprojectspring.factory.PersonnelFactory.*;
 import static org.hamcrest.Matchers.*;
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 import lombok.extern.slf4j.Slf4j;
-import ministryofeducation.sideprojectspring.factory.PersonnelFactory;
-import ministryofeducation.sideprojectspring.personnel_list.application.PersonnelListService;
-import ministryofeducation.sideprojectspring.personnel_list.presentation.PersonnelListController;
-import ministryofeducation.sideprojectspring.personnel_list.presentation.dto.response.PersonnelListDto;
-import org.hamcrest.Matchers;
+import ministryofeducation.sideprojectspring.personnel.application.PersonnelListService;
+import ministryofeducation.sideprojectspring.personnel.presentation.PersonnelListController;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
@@ -20,14 +16,11 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
